requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.103.2
flatbuffers==23.5.26
h11==0.14.0
humanfriendly==10.0
idna==3.4
imageio==2.31.5
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.3.1
jupyter_core==5.3.2
lazy_loader==0.3
llvmlite==0.41.0
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
numba==0.58.0
numpy==1.25.2
onnxruntime==1.16.0
opencv-python-headless==4.8.1.78
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
pooch==1.7.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyMatting==1.1.8
python-dateutil==2.8.2
python-multipart==0.0.6
PyWavelets==1.4.1
pyzmq==25.1.1
rembg==2.0.50
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tifffile==2023.9.26
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.23.2
wcwidth==0.2.8
